Move index unit tests into own packages
