+++ modules/core/src/test/java/org/locationtech/jts/index/strtree/GeometryDistanceComparator.java
@@ -10,7 +10,7 @@
  *
  * http://www.eclipse.org/org/documents/edl-v10.php.
  */
-package org.locationtech.jts.index;
+package org.locationtech.jts.index.strtree;
 
 import java.io.Serializable;
 import java.util.Comparator;
