+++ modules/core/src/test/java/org/locationtech/jts/index/strtree/STRtreeTest.java
@@ -10,7 +10,7 @@
  *
  * http://www.eclipse.org/org/documents/edl-v10.php.
  */
-package org.locationtech.jts.index;
+package org.locationtech.jts.index.strtree;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -18,19 +18,17 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Random;
 
+import junit.framework.TestCase;
+
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.geom.GeometryFactory;
-import org.locationtech.jts.index.strtree.AbstractNode;
-import org.locationtech.jts.index.strtree.GeometryItemDistance;
-import org.locationtech.jts.index.strtree.ItemBoundable;
-import org.locationtech.jts.index.strtree.STRtree;
+import org.locationtech.jts.geom.Point;
+import org.locationtech.jts.index.ItemVisitor;
+import org.locationtech.jts.index.SpatialIndexTester;
 import org.locationtech.jts.util.AssertionFailedException;
 
-import org.locationtech.jts.geom.Point;
-
-import junit.framework.TestCase;
 import test.jts.index.STRtreeDemo;
 import test.jts.util.SerializationUtil;
